Replacing example assets
