+++ Packages/Tracking/Examples~/XR Examples/Example Assets/Anchors.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7bc433525a2b8d043a12fa1fc634c9c2
-timeCreated: 1493926641
-licenseType: Free
+guid: 8545b8c9c65f8744fb6fdee1e4930f55
+folderAsset: yes
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
