+++ Packages/Tracking/Examples~/XR Examples/Example Assets/Anchors/scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 235a6be79d4dd9540955f0bae16a5db2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
