--- Packages/Tracking/Examples~/XR Examples/Example Assets/Physical Hands/Scripts/PhysicalHandsSimpleAnchorFeedback.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f397a96d3722f384ca1267a7d1b2544a
-timeCreated: 1494443742
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
