--- Packages/Tracking/Examples~/XR Examples/Example Assets/Physical Hands/Scripts/PhysicalHandsWorkstationBehaviourExample.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: cd5900a6b1fb91e48bd6a311753ac15f
-timeCreated: 1494460163
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
